+++ src/config.rs
@@ -90,6 +90,9 @@ pub struct Config {
     /// A format string to use to format results, similarly to exec
     pub format: Option<FormatTemplate>,
 
+    /// Whether the automatic metadata format should print a header row.
+    pub format_metadata: bool,
+
     /// If a value is supplied, each item found will be used to generate and execute commands.
     pub command: Option<Arc<CommandSet>>,